+++ frontend/src/components/control/ResumioApp/DraftSelector/DraftList.jsx
@@ -1,43 +1,15 @@
-import React, { useState } from 'react'
+import React from 'react'
 import { DraftsContainer, DraftTab } from '../../../style/ResumioApp/DraftListStyles'
-import LoadingPanel from '../../LoadingPanel'
 
-export default function DraftList({setCurrentDraft}) {
-  const getDraftsUrl = '/app/drafts'
-  const [loading, setLoading] = useState(true)
-  const [drafts, setDrafts] = useState([])
 
-  const getDraftsSuccess = (newList) => {
-    setDrafts(newList)
-    setLoading(false)
-  }
-
-  const getDraftsError = (error) => {
-    console.log('error')
-    console.log(error)
-    setLoading(false)
-  } 
-
-  if(loading){
-    return (
-      <LoadingPanel
-        method={'get'}
-        requestUrl={getDraftsUrl}
-        requestBody={{}} 
-        successCallback={getDraftsSuccess}
-        errorCallback={getDraftsError}
-      />
-    )
-  }
-
+export default function DraftList({drafts, setCurrentDraft}) {
   return (
     <DraftsContainer>
       {drafts.map((draft) => {
         return(
-          <DraftTab key={draft._id} draft={draft}>
+          <DraftTab onClick={()=>setCurrentDraft(draft)} key={draft._id} draft={draft}>
             {draft.title}
           </DraftTab>
-          
         )
       })}
     </DraftsContainer>
