--- frontend/src/components/control/ResumioApp/EmptyApp.jsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import { AppSection } from '../../style/ResumioApp/ResumioAppStyles'
-
-function EmptyApp() {
-  return (
-    <AppSection >
-      Please select an app
-    </AppSection>
-  )
-}
-
-export default EmptyApp
