+++ frontend/src/components/control/ResumioApp/InputApp/InputApp.jsx
@@ -1,24 +1,13 @@
-import React, { useState } from 'react'
+import React from 'react'
 import { NewDraftBttn } from '../../../style/ResumioApp/InputApp/InputAppStyles'
 import { AppSection } from '../../../style/ResumioApp/ResumioAppStyles'
-import DraftList from '../DraftSelector/DraftList'
 
 
-export default function InputApp() {
-  const [currentDraft, setCurrentDraft] = useState({})
 
-  const blankDraft = {
-    title: '',
-    resume: [],
-    created: '',
-    modified: ''
-  }
-  
+export default function InputApp({setCurrentDraft}) {
   return (
     <>
-      
-      <DraftList setCurrentDraft={setCurrentDraft} />
-      <NewDraftBttn onClick={()=> setCurrentDraft(blankDraft)}>+ New Draft</NewDraftBttn>
+      <NewDraftBttn onClick={()=> setCurrentDraft({})}>+ New Draft</NewDraftBttn>
       <AppSection>
         Input App
       </AppSection>
