+++ frontend/src/components/control/ResumioApp/OutputApp/OutputApp.jsx
@@ -1,13 +1,11 @@
-import React, { useState } from 'react'
+import React from 'react'
 import { AppSection } from '../../../style/ResumioApp/ResumioAppStyles'
-import DraftList from '../DraftSelector/DraftList'
+
 
 export default function OutputApp() {
-  const [currentDraft, setCurrentDraft] = useState({})
 
   return (
     <>
-      <DraftList setCurrentDraft={setCurrentDraft} />
       <AppSection>
         Output App
       </AppSection>
